Add the ability to re-evalute elite candidate.
This adds CMAParameters::set_revaluate_elite(const int &revaluate_elite)

If used, it will re-evaluate fitness value of elite candidate each
iteration. It can be useful for very noisy objective functions. In the
presence of a noisy objective function, the elite may get a fitness value
that's accidentally really high, and it will take a long time for the
rest of the population to surpass it. In some sense, that elite is
really a fake elite; it only got the high fitness score by accident.

This changeset makes it so that the elite re-computation is done in the
same OpenMP block as the rest of the candidate evaluations to make sure
it gets the same multithreaded treatment.